requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
openai>=2.24.0
python-dotenv==1.0.1
tenacity==9.0.0
jinja2==3.1.4
